pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Lukash <89802614+AlexanderLukash@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.1.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
